requirements.txt
asgiref==3.4.1
Django==3.0.10
django-cors-headers==3.7.0
django-mssql-backend==2.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
greenlet==1.1.1
numpy==1.21.1
openpyxl==3.0.7
ordered-set==4.0.2
pandas==1.3.1
ply==3.11
PyJWT==2.1.0
pyodbc==4.0.32
Pyomo==6.0.1
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
ticdat==0.2.21
whitenoise==5.3.0